requirements.txt
nltk
Flask==3.0.2
passlib==1.7.4
pydantic==2.6.4
PyJWT==2.8.0
PyJWT==2.8.0
pymongo==4.6.2
python-dotenv==1.0.1
Werkzeug==3.0.1
waitress
requests